Create Attribute, ComboHelper and Enums files and move relevant functions from TestValues into them
