+++ Source/AssetGenerator.csproj
@@ -49,9 +49,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Attribute.cs" />
     <Compile Include="Attributes.cs" />
+    <Compile Include="ComboHelper.cs" />
     <Compile Include="Common.cs" />
     <Compile Include="Data.cs" />
+    <Compile Include="Enums.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Runtime\GLTF.cs" />
