+++ Source/ComboHelper.cs
@@ -1,0 +1,344 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace AssetGenerator
+{
+    public static class ComboHelper
+    {
+        public static List<List<Attribute>> AttributeCombos(TestValues test)
+        {
+            List<List<Attribute>> finalResult;
+            List<List<Attribute>> removeTheseCombos = new List<List<Attribute>>();
+            List<List<Attribute>> keepTheseCombos = new List<List<Attribute>>();
+            List<Attribute> isRequired = new List<Attribute>();
+            List<Attribute> isPrerequisite = new List<Attribute>();
+            bool hasPrerequisiteAttribute;
+
+            //var combos = PowerSet<Attribute>(attributes);
+            var combos = BasicSet<Attribute>(test.attributes);
+
+            // Include any special combos
+            if (test.specialCombos.Any())
+            {
+                foreach (var x in test.specialCombos)
+                {
+                    var comboIndex = combos.FindIndex(e => e.Any() && e[0].name == x[0].name && e.Count() == 1);
+                    combos.Insert(comboIndex + 1, x);
+                }
+            }
+
+            // Remove the explicitly excluded combos
+            if (test.removeCombos.Any())
+            {
+                foreach (var x in test.removeCombos)
+                {
+                    combos.RemoveAll(e => e.Count == 1 && e[0].name == x[0].name);
+                }
+            }
+
+            if (test.noPrerequisite == false)
+            {
+                // Makes a list of names of possible prerequisites
+                List<AttributeName> Prerequisites = new List<AttributeName>();
+                foreach (var x in test.attributes)
+                {
+                    if (x.prerequisite != AttributeName.Undefined)
+                    {
+                        if (Prerequisites.Any())
+                        {
+                            bool isNew = true;
+                            foreach (var y in Prerequisites)
+                            {
+                                if (y == x.prerequisite)
+                                {
+                                    isNew = false;
+                                    break;
+                                }
+                            }
+                            if (isNew == true)
+                            {
+                                Prerequisites.Add(x.prerequisite);
+                            }
+                        }
+                        else
+                        {
+                            Prerequisites.Add(x.prerequisite);
+                        }
+                    }
+                }
+                // Convert the name list into a list of the actual prerequisites
+                foreach (var x in Prerequisites)
+                {
+                    foreach (var y in test.attributes)
+                    {
+                        if (x == y.name)
+                        {
+                            isPrerequisite.Add(y);
+                            break;
+                        }
+                    }
+                }
+                // Add combos where prerequisite attributes have all dependant attributes set
+                foreach (var x in isPrerequisite)
+                {
+                    // Start a list with the prerequisite attribute 
+                    var addList = new List<Attribute>
+                    {
+                        x
+                    };
+
+                    // Populate that list will all of the required attributes
+                    foreach (var y in test.attributes)
+                    {
+                        if (y.prerequisite == x.name)
+                        {
+                            addList.Add(y);
+                        }
+                    }
+                    // Then include the combo with the rest
+                    var comboIndex = combos.FindIndex(e => e.Any() && e[0].name == addList[0].name && e.Count() == 1);
+                    combos.Insert(comboIndex + 1, addList);
+                }
+            }
+
+            // Handle non-binary attributes in the first combo
+            if (test.onlyBinaryAttributes == false)
+            {
+                List<Attribute> keep = new List<Attribute>();
+                foreach (var x in combos[1])
+                {
+                    // Keep attribute if it is the first found or is binary
+                    if (x.attributeGroup == 0 || (x.attributeGroup > 0 && !keep.Any()))
+                    {
+                        keep.Add(x);
+                    }
+                    else if (x.attributeGroup > 0)
+                    {
+                        bool alreadyKept = false;
+                        foreach (var y in keep)
+                        {
+                            // Don't keep the nonbinary attribute if there is already one of that set on the list
+                            if (y.attributeGroup == x.attributeGroup)
+                            {
+                                alreadyKept = true;
+                                break;
+                            }
+                        }
+                        if (alreadyKept == false) // Keep nonbinary attribute 
+                        {
+                            keep.Add(x);
+                        }
+                    }
+                }
+                // Remove the extra nonbinary attributes
+                combos[1] = keep;
+            }
+
+            // Removes sets that duplicate binary entries for a single attribute (e.g. alphaMode)
+            // Removes sets where an attribute is missing a required attribute
+            if (test.onlyBinaryAttributes == false || test.noPrerequisite == false)
+            {
+                // Are there any prerequisite attributes? 
+                hasPrerequisiteAttribute = isPrerequisite.Any();
+
+                // Makes a list of combos to remove
+                int combosCount = combos.Count();
+                for (int x = 2; x < combosCount; x++) // The first two combos are already taken care of
+                {
+                    bool usedPrereq = false;
+                    List<int> binarySets = new List<int>();
+                    List<AttributeName> usedPrerequisite = new List<AttributeName>();
+
+                    // Makes a list of each prerequisite attribute in the current combo
+                    if (hasPrerequisiteAttribute == true)
+                    {
+                        foreach (var prereq in isPrerequisite)
+                        {
+                            foreach (var attribute in combos[x])
+                            {
+                                if (attribute.name == prereq.name)
+                                {
+                                    usedPrerequisite.Add(prereq.name);
+                                }
+                            }
+                        }
+                        usedPrereq = usedPrerequisite.Any();
+                    }
+
+                    foreach (var attribute in combos[x])
+                    {
+                        // Remove combos that have multiple of the same binary combo
+                        if (attribute.attributeGroup > 0)
+                        {
+                            if (binarySets.Contains(attribute.attributeGroup))
+                            {
+                                removeTheseCombos.Add(combos[x]);
+                                break;
+                            }
+                            else
+                            {
+                                binarySets.Add(attribute.attributeGroup);
+                            }
+                        }
+                        // Removes combos that have a attribute missing a prerequisite
+                        if (usedPrereq == true && attribute.prerequisite != AttributeName.Undefined)
+                        {
+                            bool prereqNotFound = true;
+                            foreach (var prereq in usedPrerequisite)
+                            {
+                                if (attribute.prerequisite == prereq)
+                                {
+                                    prereqNotFound = false;
+                                    break;
+                                }
+                            }
+                            if (prereqNotFound != false)
+                            {
+                                removeTheseCombos.Add(combos[x]);
+                                break;
+                            }
+                        }
+                        else if (usedPrereq == false && attribute.prerequisite != AttributeName.Undefined)
+                        {
+                            removeTheseCombos.Add(combos[x]);
+                            break;
+                        }
+                    }
+                }
+
+                // Uses the list of bad combos to trim down the original set
+                int numCombos = combos.Count();
+                int numRemoveTheseCombos = removeTheseCombos.Count();
+                for (int x = 0; x < numCombos; x++)
+                {
+                    bool excludeCombo = false;
+                    for (int y = 0; y < numRemoveTheseCombos; y++)
+                    {
+                        if (combos[x] == removeTheseCombos[y])
+                        {
+                            excludeCombo = true;
+                            break;
+                        }
+                    }
+                    if (excludeCombo == false)
+                    {
+                        keepTheseCombos.Add(combos[x]);
+                    }
+                }
+                finalResult = keepTheseCombos;
+            }
+            else
+            {
+                // If there are only binary attributes, we don't need to check for duplicates
+                finalResult = combos;
+            }
+
+            return finalResult;
+        }
+
+        public static List<Attribute> CustomComboCreation(Attribute attributeA, Attribute attributeB = null, Attribute attributeC = null)
+        {
+            List<Attribute> newCombo = new List<Attribute>();
+
+            newCombo.Add(attributeA);
+
+            if (attributeB != null)
+            {
+                newCombo.Add(attributeB);
+            }
+
+            if (attributeC != null)
+            {
+                newCombo.Add(attributeC);
+            }
+
+            return newCombo;
+        }
+
+        /// <summary>
+        /// Given a,b,c this returns [a,b,c], [a], [b], [c]
+        /// </summary>
+        /// <param name="seq"></param>
+        /// <returns>List of lists containing a basic set of combinations.</returns>
+        public static List<List<T>> BasicSet<T>(List<T> seq)
+        {
+            var basicSet = new List<List<T>>();
+            basicSet.Add(new List<T>()); // Will contain the empty set
+            basicSet.Add(new List<T>()); // Will contain the full set
+            foreach (var x in seq)
+            {
+                basicSet[1].Add(x);
+                var addList = new List<T>
+                {
+                    x
+                };
+                basicSet.Add(addList);
+            }
+            return basicSet;
+        }
+
+        /// <summary>
+        /// Given a,b,c this returns all possible combinations including a full and empty set.
+        /// </summary>
+        /// <param name="seq"></param>
+        /// <returns>List of lists containing a powerset.</returns>
+        //https://stackoverflow.com/questions/19890781/creating-a-power-set-of-a-sequence
+        public static List<List<T>> PowerSet<T>(List<T> seq)
+        {
+            var powerSet = new List<List<T>>();
+            powerSet.Add(new List<T>()); // starting only with empty set
+            for (int i = 0; i < seq.Count; i++)
+            {
+                var cur = seq[i];
+                int count = 1 << i; // doubling list each time
+                for (int j = 0; j < count; j++)
+                {
+                    var source = powerSet[j];
+                    powerSet.Add(new List<T>());
+                    powerSet[count + j] = new List<T>();
+                    var destination = powerSet[count + j];
+                    for (int q = 0; q < source.Count; q++)
+                        destination.Add(source[q]);
+                    destination.Add(cur);
+                }
+            }
+            return powerSet;
+        }
+
+        /// <summary>
+        /// Compares two sets of model attributes and returns whether they are equal or not, regardless of list order
+        /// </summary>
+        /// <param name="comboToCheck"></param>
+        /// <param name="comboToFind"></param>
+        /// <returns>Returns a bool, true if they contain the exact same attributes in any order</returns>
+        public static bool FindCombo(Attribute[] comboToCheck, Attribute[] comboToFind)
+        {
+            if (comboToCheck.Count() == comboToFind.Count())
+            {
+                bool isEqual = true;
+                foreach (var x in comboToCheck)
+                {
+                    bool containsElement = false;
+                    foreach (var y in comboToFind)
+                    {
+                        if (x.name == y.name)
+                        {
+                            containsElement = true;
+                            break;
+                        }
+                    }
+                    if (containsElement == false)
+                    {
+                        isEqual = false;
+                        break;
+                    }
+                }
+                if (isEqual == true)
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+    }
+}
