+++ Source/Enums.cs
@@ -1,0 +1,65 @@
+﻿namespace AssetGenerator
+{    public enum Tests
+    {
+        Undefined,
+        Material,
+        Material_Alpha,
+        Material_MetallicRoughness,
+        Texture_Sampler,
+        Primitive_Attribute,
+    }
+    public enum AttributeName
+    {
+        Undefined,
+        Name,
+        BaseColorFactor,
+        BaseColorTexture,
+        MetallicFactor,
+        RoughnessFactor,
+        MetallicRoughnessTexture,
+        PbrTextures,
+        EmissiveFactor,
+        AlphaMode_Mask,
+        AlphaMode_Blend,
+        AlphaMode_Opaque,
+        AlphaCutoff,
+        Color_Vector3_Float,
+        Color_Vector4_Float,
+        Color_Vector3_Byte,
+        Color_Vector4_Byte,
+        Color_Vector3_Short,
+        Color_Vector4_Short,
+        DoubleSided,
+        Sampler,
+        MagFilter_Nearest,
+        MagFilter_Linear,
+        MinFilter_Nearest,
+        MinFilter_Linear,
+        MinFilter_NearestMipmapNearest,
+        MinFilter_LinearMipmapNearest,
+        MinFilter_NearestMipmapLinear,
+        MinFilter_LinearMipmapLinear,
+        Normal,
+        Position,
+        Tangent,
+        TexCoord0_Float,
+        TexCoord0_Byte,
+        TexCoord0_Short,
+        TexCoord1_Float,
+        TexCoord1_Byte,
+        TexCoord1_Short,
+        WrapS_ClampToEdge,
+        WrapS_MirroredRepeat,
+        WrapS_Repeat,
+        WrapT_ClampToEdge,
+        WrapT_MirroredRepeat,
+        WrapT_Repeat,
+        Source,
+        TexCoord,
+        NormalTexture,
+        OcclusionTexture,
+        EmissiveTexture,
+        Scale,
+        Strength
+    }
+}
